Open nodes close to a particle softening length slightly more often.
Open a node when some but not all of it is within a force softening
length of particle. When all the node is inside, all is softened and
there is no need to open it.

When checking MixedSoftenings, which we do because we cannot tell how
much of the node actually has this softening, check intersection
with the total node instead of just the center.

It should be rare that these conditions are satisfied without the
particle being within the node anyway. This needs nop->len to be small,
in which case h + nop->len/2 ~ h anyway.